updated x509/x509.go to stdlib porting changes to x509/x509_modified.go
